move test setup files
