+++ test/e2e/features/data-explorer/data-explorer-headless.test.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { join } from 'path';
-import { test, expect } from '../../_test.setup';
+import { test, expect } from '../_test.setup';
 import { Application, Logger } from '../../../automation';
 
 test.use({
