+++ test/e2e/features/data-explorer/data-explorer-python-pandas.test.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { join } from 'path';
-import { test, expect } from '../../_test.setup';
+import { test, expect } from '../_test.setup';
 
 test.use({
 	suiteId: __filename
