+++ test/e2e/features/data-explorer/very-large-data-frame.test.ts
@@ -6,7 +6,7 @@
 import { fail } from 'assert';
 import { join } from 'path';
 import { downloadFileFromS3, S3FileDownloadOptions } from '../../../automation';
-import { test, expect } from '../../_test.setup';
+import { test, expect } from '../_test.setup';
 
 test.use({
 	suiteId: __filename
