+++ test/e2e/features/new-project-wizard/new-project-python.test.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { PositronPythonFixtures, ProjectType, ProjectWizardNavigateAction } from '../../../automation';
-import { test, expect } from '../../_test.setup';
+import { test, expect } from '../_test.setup';
 
 test.use({
 	suiteId: __filename
