+++ test/e2e/features/new-project-wizard/new-project-r-jupyter.test.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { ProjectType, ProjectWizardNavigateAction } from '../../../automation';
-import { test, expect } from '../../_test.setup';
+import { test, expect } from '../_test.setup';
 
 test.use({
 	suiteId: __filename
